# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      task_average_completion_time:
        type: string
      tasks_completed:
        type: integer
      tasks_completed_rate:
        type: number
      tasks_created:
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      id:
        type: integer
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-bool'
      description:
        $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-string'
      title:
        $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-string'
      phone_number:
        $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      version:
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo application REST API scheme
  title: Todo API
  version: "1.0"
paths:
  /statistics:
    get:
      description: Get statistics for tasks created and completed in a specified date
        range. If user_id is not provided, returns statistics for all users.
      parameters:
      - description: User ID to get statistics for (optional, returns all users if
          omitted)
        in: query
        name: user_id
        type: integer
      - description: Start date in ISO 8601 format (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date in ISO 8601 format (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Statistics retrieved successfully
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad Request - invalid query parameters
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Get task statistics
      tags:
      - statistics
  /tasks:
    get:
      description: List all tasks with optional pagination
      parameters:
      - description: limit of page
        in: query
        name: limit
        type: integer
      - description: offset of page
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Task Deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Delete task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      parameters:
      - description: task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch task request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Patch task
      tags:
      - tasks
  /users:
    get:
      description: List all users with optional pagination
      parameters:
      - description: limit of page
        in: query
        name: limit
        type: integer
      - description: offset of page
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: Users ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: User Deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: Users ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch user request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Censtro_todo-api_internal_core_transport_http_response.ErrorResponse'
      summary: Patch user
      tags:
      - users
swagger: "2.0"